Moved keystore to librekey
