+++ include/rnp_def.h
@@ -89,4 +89,6 @@ enum {
 
 };
 
+enum { MAX_ID_LENGTH = 128, MAX_PASSPHRASE_LENGTH = 256 };
+
 #endif
