+++ src/lib/Makefile.am
@@ -4,7 +4,7 @@ AM_CFLAGS			= $(WARNCFLAGS) $(BOTAN_INCLUDES) $(JSONC_INCLUDES)
 AM_LDFLAGS          = $(BOTAN_LDFLAGS) $(JSONC_LDFLAGS)
 lib_LTLIBRARIES		= librnp.la
 librnp_la_CPPFLAGS	= -I$(top_srcdir)/include
-librnp_la_LIBADD	= $(BOTAN_LIBS) $(JSONC_LIBS)
+librnp_la_LIBADD	= $(BOTAN_LIBS) $(JSONC_LIBS) ../librekey/librekey.la
 dist_man_MANS		= librnp.3
 librnp_la_SOURCES	= \
 	bn.c \
@@ -16,10 +16,6 @@ librnp_la_SOURCES	= \
 	eddsa.c \
 	elgamal.c \
 	hash.c \
-	key_store.c \
-	key_store_pgp.c \
-	key_store_kbx.c \
-	key_store_ssh.c \
 	list.c \
 	misc.c \
 	packet-create.c \
