+++ src/lib/crypto.h
@@ -58,7 +58,6 @@
 #include <limits.h>
 #include <botan/ffi.h>
 #include "hash.h"
-#include "key_store_pgp.h"
 #include "packet.h"
 #include "memory.h"
 #include "packet-parse.h"
