+++ src/lib/reader.c
@@ -114,7 +114,6 @@ __RCSID("$NetBSD: reader.c,v 1.49 2012/03/05 02:20:18 christos Exp $");
 #include "packet-parse.h"
 #include "packet-show.h"
 #include "packet-print.h"
-#include "key_store_pgp.h"
 #include "readerwriter.h"
 #include "rnpsdk.h"
 #include "utils.h"
