+++ src/lib/writer.c
@@ -75,7 +75,6 @@ __RCSID("$NetBSD: writer.c,v 1.33 2012/03/05 02:20:18 christos Exp $");
 #include "bn.h"
 #include "packet-create.h"
 #include "writer.h"
-#include "key_store_pgp.h"
 #include "signature.h"
 #include "packet.h"
 #include "packet-parse.h"
