+++ src/librekey/key_store_ssh.h
@@ -1,6 +1,5 @@
 /*
  * Copyright (c) 2017, [Ribose Inc](https://www.ribose.com).
- * Copyright (c) 2009 The NetBSD Foundation, Inc.
  * All rights reserved.
  *
  * This code is originally derived from software contributed to
@@ -28,46 +27,19 @@
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
  */
-/*
- * Copyright (c) 2005-2008 Nominet UK (www.nic.uk)
- * All rights reserved.
- * Contributors: Ben Laurie, Rachel Willmer. The Contributors have asserted
- * their moral rights under the UK Copyright Design and Patents Act 1988 to
- * be recorded as the authors of this copyright work.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not
- * use this file except in compliance with the License.
- *
- * You may obtain a copy of the License at
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
 
-/** \file
- */
+#ifndef KEY_STORE_SSH_H_
+#define KEY_STORE_SSH_H_
 
-#ifndef KEY_STORE_PGP_H_
-#define KEY_STORE_PGP_H_
+#include "rnp.h"
+#include <rnp_key_store.h>
 
-#include "packet.h"
-#include "packet-parse.h"
-#include "key_store.h"
-#include "memory.h"
+int rnp_key_store_ssh_load_keys(rnp_t *rnp, const char *pubpath, const char *secpath);
 
-enum { MAX_ID_LENGTH = 128, MAX_PASSPHRASE_LENGTH = 256 };
+int rnp_key_store_ssh_from_file(pgp_io_t *, rnp_key_store_t *, const char *);
+int rnp_key_store_ssh_from_mem(pgp_io_t *, rnp_key_store_t *, pgp_memory_t *);
 
-bool rnp_key_store_pgp_read_from_mem(pgp_io_t *,
-                                     rnp_key_store_t *,
-                                     const unsigned,
-                                     pgp_memory_t *);
+int rnp_key_store_ssh_to_file(pgp_io_t *, rnp_key_store_t *, const uint8_t *, const char *);
+int rnp_key_store_ssh_to_mem(pgp_io_t *, rnp_key_store_t *, const uint8_t *, pgp_memory_t *);
 
-int rnp_key_store_pgp_write_to_mem(
-  pgp_io_t *, rnp_key_store_t *, const uint8_t *, const unsigned, pgp_memory_t *);
-
-#endif /* KEY_STORE_PGP_H_ */
+#endif /* KEY_STORE_SSH_H_ */
