+++ src/tests/exportkey.c
@@ -26,7 +26,7 @@
  */
 
 #include <rnp.h>
-#include <key_store.h>
+#include <rnp_key_store.h>
 #include "rnp_tests.h"
 #include "support.h"
