+++ src/tests/pgp-parse.c
@@ -29,7 +29,6 @@
 #include <eddsa.h>
 #include <elgamal.h>
 #include <crypto.h>
-#include <key_store_pgp.h>
 #include <packet.h>
 #include <pgp-key.h>
 #include <bn.h>
